hf_webapp2/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch==2.4.0+cpu
torchvision==0.19.0+cpu
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
tqdm==4.66.5
requests
